test: docs 작업을 위한 class 분리
