+++ module-api/src/test/java/com/yapp/sharefood/common/PreprocessController.java
@@ -1,19 +1,11 @@
 package com.yapp.sharefood.common;
 
 import com.yapp.sharefood.auth.token.TokenProvider;
-import com.yapp.sharefood.common.documentation.DocumentRequestBuilder;
 import com.yapp.sharefood.user.domain.OAuthType;
 import com.yapp.sharefood.user.domain.User;
 import com.yapp.sharefood.user.repository.UserRepository;
-import org.junit.Before;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.restdocs.RestDocumentationExtension;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
 
 import java.util.Optional;
 
@@ -21,29 +13,19 @@ import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.willReturn;
 
-@ExtendWith({SpringExtension.class, RestDocumentationExtension.class})
-@AutoConfigureRestDocs
-public abstract class PreprocessController {
+public abstract class PreprocessController extends DocumentTest{
 
     @MockBean
     protected UserRepository userRepository;
     @MockBean
     protected TokenProvider tokenProvider;
 
-    @Autowired
-    protected MockMvc mockMvc;
-
-    private DocumentRequestBuilder documentRequestBuilder;
-
     protected User user;
 
     protected Long authUserId = 1L;
 
     @BeforeEach
-    void setUp() {
-
-        documentRequestBuilder = new DocumentRequestBuilder();
-
+    void setUpPreprocess() {
         user = User.builder()
                 .id(1L)
                 .nickname("nickname")
@@ -59,9 +41,5 @@ public abstract class PreprocessController {
 
         willReturn(Optional.of(user))
                 .given(userRepository).findById(anyLong());
-    }
-
-    protected DocumentRequestBuilder.MockMvcFunction document() {
-        return documentRequestBuilder.build().mockMvc(this.mockMvc);
     }
 }
